Renamed sentiment-analysis to sentimentanalysis to get around a syntax error that arises when trying to import the analyze sentence function. Also modified test file to use pytest and simplified the wording of the tests.
